requirements.txt
Django>=5.0
django-ckeditor>=6.0
